+++ modules/visual-tests/src/routes/test/Layout.tsx
@@ -17,12 +17,21 @@ const toggle = (on: true | undefined) => !on || undefined;
 // Use `handle` if you're using multiple Hanna compnents
 export const handle = cssTokens('Alert', 'BreadCrumbs');
 
+const siteNames = [
+  'Mínar síður',
+  'Gagnahlaðborðið í tveimur línum',
+  'YfirgengilegaLangtHeiti',
+];
+
 export default function () {
   const [globalAlerts, setGlobalAlerts] = useState<true | undefined>();
   const [bigContent, setBigContent] = useState<true | undefined>();
+  const [siteName, setSiteName] = useState<string | undefined>();
+  const [opacity, setOpacity] = useState(0);
 
   return (
     <Layout
+      siteName={siteName}
       globalAlerts={
         globalAlerts && (
           <Fragment>
@@ -37,7 +46,7 @@ export default function () {
       }
       navChildren={
         <Fragment>
-          <BreadCrumbs trail={breadCrumbTrail} />
+          {!siteName && <BreadCrumbs trail={breadCrumbTrail} />}
           <DummyBlock thin />
           Nav Content
           <DummyBlock thin />
@@ -50,6 +59,10 @@ export default function () {
           <DummyBlock thin />
         </Fragment>
       }
+      wrapperProps={{
+        onMouseEnter: () => setOpacity(1),
+        onMouseLeave: () => setOpacity(0),
+      }}
     >
       <DummyBlock thin />
       <p>
@@ -57,9 +70,15 @@ export default function () {
         <button data-testid="content" onClick={() => setBigContent(toggle)}>
           BigContent
         </button>
-        ,{' '}
         <button data-testid="alerts" onClick={() => setGlobalAlerts(toggle)}>
-          Alerts
+          , Alerts
+        </button>
+        <button
+          data-testid="sitename"
+          onClick={() => setSiteName((n) => siteNames[siteNames.indexOf(n || '') + 1])}
+          style={{ opacity }}
+        >
+          , SiteName
         </button>
       </p>
       {bigContent ? <DummyBlock customHeight="600px" /> : <DummyBlock big />}
@@ -68,14 +87,27 @@ export default function () {
 }
 
 export const testing: TestingInfo = {
-  extras: async ({ page, pageScreenshot }) => {
+  extras: async ({ page, pageScreenshot, localScreenshot, mediaFormat }) => {
     await page.getByTestId('content').click();
     await page.mouse.move(0, 0);
     await pageScreenshot('bigcontent');
-
     await page.getByTestId('content').click();
+
     await page.getByTestId('alerts').click();
     await page.mouse.move(0, 0);
     await pageScreenshot('alerts');
+    await page.getByTestId('alerts').click();
+
+    if (mediaFormat('wide') || mediaFormat('phone')) {
+      const homeLink = page.locator('.Layout__header__homelink');
+      let i = siteNames.length;
+      while (i--) {
+        await page.locator('.Layout__main').hover();
+        await page.getByTestId('sitename').click();
+        await localScreenshot(homeLink, `siteneame-${i}`, { margin: true });
+      }
+      await page.getByTestId('sitename').click();
+      await page.mouse.move(0, 0);
+    }
   },
 };