+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `Layout` and `WizardLayout`:
+  - feat: Add `__header__homelink` and `__header__sitename`
 - feat: Add component `AutosuggestSearch`
 - feat: Make `SiteSearchInput`'s `<button/>` optional
 - feat: Add component `Pagination`